test(server): finish splitting lib tests by topic
