+++ nodelite-server/src/tests/sanitize_snapshot_tests.rs
@@ -1,15 +1,10 @@
-use std::net::{IpAddr, Ipv4Addr, SocketAddr, SocketAddrV4};
+//! Snapshot sanitization tests.
+
+use std::net::{Ipv4Addr, SocketAddr, SocketAddrV4};
 use std::path::PathBuf;
 
-use axum::http::HeaderMap;
 use chrono::Utc;
 
-use super::support::trusted_proxies;
-use crate::admission::{
-    InstallAdmissionConfig, InstallAdmissionController, WsAdmissionController, WsAdmissionError,
-    resolve_client_ip, sweep_expired_auth_failures,
-};
-use crate::handlers::is_well_formed_install_token;
 use crate::sanitize::{
     MAX_SANITIZED_DISKS, MAX_SANITIZED_LOAD, MAX_SANITIZED_RATE_BYTES_PER_SEC,
     MAX_SANITIZED_STRING_BYTES, METRIC_ANOMALY_SESSION_LIMIT, SanitizationReport,
@@ -353,7 +348,6 @@ fn sanitize_snapshot_caps_disk_count_and_tracks_clean_reports() {
     assert_eq!(report.dropped_disks, 3);
     assert!(report.modified());
 
-    // clean 报告不应推动 anomaly 窗口前进;modified 报告才入窗口。
     let mut window: std::collections::VecDeque<std::time::Instant> =
         std::collections::VecDeque::new();
     let now = std::time::Instant::now();
@@ -361,7 +355,6 @@ fn sanitize_snapshot_caps_disk_count_and_tracks_clean_reports() {
     update_metric_anomaly_window(&mut window, &clean_report, now);
     assert!(window.is_empty());
 
-    // 在窗口内攒满 METRIC_ANOMALY_SESSION_LIMIT 条 → 触发断连。
     for tick in 0..METRIC_ANOMALY_SESSION_LIMIT {
         update_metric_anomaly_window(
             &mut window,
@@ -478,14 +471,12 @@ fn sanitize_snapshot_deduplicates_repeated_disk_devices() {
 
 #[test]
 fn truncate_to_byte_boundary_respects_char_boundary() {
-    // "中" 在 UTF-8 中占 3 字节;cutoff = 7 必须回退到 6 字节边界。
     let mut value = "中".repeat(100);
     nodelite_proto::truncate_string_to_byte_boundary(&mut value, 7);
     assert!(value.len() <= 7);
     assert!(value.is_char_boundary(value.len()));
     assert!(value.chars().all(|ch| ch == '中'));
 
-    // 已经在限内的字符串保持不变。
     let mut short = "abc".to_string();
     nodelite_proto::truncate_string_to_byte_boundary(&mut short, 16);
     assert_eq!(short, "abc");
@@ -507,266 +498,5 @@ fn truncate_to_byte_boundary_handles_utf8_widths_with_bounded_scan() {
         assert_eq!(value, expected);
         assert!(value.len() <= max_bytes);
         assert!(value.is_char_boundary(value.len()));
-    }
-}
-
-#[test]
-fn loopback_proxy_peer_uses_forwarded_ip_for_ws_limits() {
-    let mut headers = HeaderMap::new();
-    headers.insert(
-        "x-forwarded-for",
-        "198.51.100.24".parse().expect("header value"),
-    );
-
-    let client_ip = resolve_client_ip(
-        &[],
-        SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 51234)),
-        &headers,
-    );
-
-    assert_eq!(client_ip, IpAddr::V4("198.51.100.24".parse().expect("ip")));
-}
-
-#[test]
-fn public_listener_behind_local_proxy_uses_forwarded_ip() {
-    let mut headers = HeaderMap::new();
-    headers.insert(
-        "x-forwarded-for",
-        "198.51.100.24".parse().expect("header value"),
-    );
-
-    let client_ip = resolve_client_ip(
-        &[],
-        SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 51234)),
-        &headers,
-    );
-
-    assert_eq!(client_ip, IpAddr::V4("198.51.100.24".parse().expect("ip")));
-}
-
-#[test]
-fn public_direct_peer_ignores_spoofed_forwarded_ip() {
-    let mut headers = HeaderMap::new();
-    headers.insert("x-forwarded-for", "8.8.8.8".parse().expect("header value"));
-
-    let client_ip = resolve_client_ip(
-        &[],
-        SocketAddr::V4(SocketAddrV4::new(
-            "198.51.100.24".parse().expect("ip"),
-            51234,
-        )),
-        &headers,
-    );
-
-    assert_eq!(client_ip, IpAddr::V4("198.51.100.24".parse().expect("ip")));
-}
-
-#[test]
-fn trusted_proxy_chain_uses_last_untrusted_forwarded_ip() {
-    let mut headers = HeaderMap::new();
-    headers.insert(
-        "x-forwarded-for",
-        "8.8.8.8, 198.51.100.24, 203.0.113.11"
-            .parse()
-            .expect("header value"),
-    );
-
-    let client_ip = resolve_client_ip(
-        &trusted_proxies(&["203.0.113.0/24"]),
-        SocketAddr::V4(SocketAddrV4::new(
-            "203.0.113.10".parse().expect("ip"),
-            51234,
-        )),
-        &headers,
-    );
-
-    assert_eq!(client_ip, IpAddr::V4("198.51.100.24".parse().expect("ip")));
-}
-
-#[test]
-fn trusted_proxy_prefers_x_real_ip_when_forwarded_chain_is_absent() {
-    let mut headers = HeaderMap::new();
-    headers.insert("x-real-ip", "198.51.100.24".parse().expect("header value"));
-
-    let client_ip = resolve_client_ip(
-        &trusted_proxies(&["203.0.113.0/24"]),
-        SocketAddr::V4(SocketAddrV4::new(
-            "203.0.113.10".parse().expect("ip"),
-            51234,
-        )),
-        &headers,
-    );
-
-    assert_eq!(client_ip, IpAddr::V4("198.51.100.24".parse().expect("ip")));
-}
-
-#[test]
-fn malformed_forwarded_chain_falls_back_to_x_real_ip() {
-    let mut headers = HeaderMap::new();
-    headers.insert(
-        "x-forwarded-for",
-        "8.8.8.8, invalid-ip".parse().expect("header value"),
-    );
-    headers.insert("x-real-ip", "198.51.100.24".parse().expect("header value"));
-
-    let client_ip = resolve_client_ip(
-        &[],
-        SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 51234)),
-        &headers,
-    );
-
-    assert_eq!(client_ip, IpAddr::V4("198.51.100.24".parse().expect("ip")));
-}
-
-#[test]
-fn repeated_auth_failures_trigger_ws_block() {
-    let controller = WsAdmissionController::new(&WsConfig {
-        max_total_connections: 16,
-        max_connections_per_ip: 4,
-        auth_fail_window_secs: 60,
-        auth_fail_max_attempts: 2,
-        auth_block_secs: 300,
-    });
-    let client_ip = IpAddr::V4("198.51.100.24".parse().expect("ip"));
-
-    controller.record_auth_failure(client_ip);
-    controller.record_auth_failure(client_ip);
-
-    match controller.try_acquire(client_ip) {
-        Err(WsAdmissionError::Blocked { retry_after_secs }) => {
-            assert!(retry_after_secs > 0);
-        }
-        _ => panic!("client should be temporarily blocked"),
-    }
-}
-
-#[test]
-fn metric_anomaly_window_decays_so_long_sessions_avoid_false_positive_kicks() {
-    // 旧实现:METRIC_ANOMALY_SESSION_LIMIT 是会话生命周期内的累计上限,
-    // 因此长跑节点偶发 5 次异常就会被踢。
-    // 新实现:计数滑动到 METRIC_ANOMALY_WINDOW_SECS 之外即衰减,只有
-    // "在同一窗口内连续超阈值"才触发断连。
-    use std::collections::VecDeque;
-    use std::time::{Duration, Instant};
-
-    let mut window: VecDeque<Instant> = VecDeque::new();
-    let report = SanitizationReport {
-        clamped_percents: 1,
-        ..SanitizationReport::default()
-    };
-
-    // 模拟一个 24 小时的长会话,每隔 1 小时遇到一次偶发的 sanitize 修正。
-    // 任何两次 anomaly 的间隔(3600 s)都远大于窗口长度(默认 300 s),
-    // 因此每次入队前老条目都已被剔除,窗口始终最多只有 1 条。
-    let started_at = Instant::now();
-    for hour in 0..24 {
-        let now = started_at + Duration::from_secs(hour * 3600);
-        update_metric_anomaly_window(&mut window, &report, now);
-        assert!(
-            !should_disconnect_for_metric_anomalies(&window),
-            "long session with sparse anomalies should never be kicked",
-        );
-    }
-
-    // 反过来,同一窗口内的高频异常 → 窗口内累计达到阈值 → 触发断连。
-    let burst_at = started_at + Duration::from_secs(48 * 3600);
-    for tick in 0..METRIC_ANOMALY_SESSION_LIMIT {
-        update_metric_anomaly_window(
-            &mut window,
-            &report,
-            burst_at + Duration::from_secs(tick as u64),
-        );
-    }
-    assert!(
-        should_disconnect_for_metric_anomalies(&window),
-        "burst within the window must still trigger the kick",
-    );
-}
-
-#[test]
-fn sweep_drops_expired_failure_entries_and_keeps_live_blocks() {
-    // 验证 sweep:已过期且未封禁的条目被移除;仍封禁的条目保留;
-    // 仍在统计窗口内的失败条目保留。
-    use std::collections::{HashMap, VecDeque};
-    use std::time::{Duration, Instant};
-
-    use crate::admission::AuthFailureState;
-
-    let mut failures: HashMap<IpAddr, AuthFailureState> = HashMap::new();
-    let now = Instant::now();
-    let window = Duration::from_secs(60);
-
-    // 1. 过期 + 未封禁 → 应被 sweep 删除
-    let expired_ip: IpAddr = "203.0.113.10".parse().expect("ip");
-    let mut expired = AuthFailureState::default();
-    expired
-        .recent_failures
-        .push_back(now - Duration::from_secs(3600));
-    failures.insert(expired_ip, expired);
-
-    // 2. 已封禁但封禁未到期 → 应保留
-    let blocked_ip: IpAddr = "203.0.113.20".parse().expect("ip");
-    let blocked = AuthFailureState {
-        recent_failures: VecDeque::new(),
-        blocked_until: Some(now + Duration::from_secs(300)),
-    };
-    failures.insert(blocked_ip, blocked);
-
-    // 3. 窗口内的失败 → 应保留
-    let recent_ip: IpAddr = "203.0.113.30".parse().expect("ip");
-    let mut recent = AuthFailureState::default();
-    recent
-        .recent_failures
-        .push_back(now - Duration::from_secs(10));
-    failures.insert(recent_ip, recent);
-
-    sweep_expired_auth_failures(&mut failures, now, window);
-
-    assert!(
-        !failures.contains_key(&expired_ip),
-        "expired entry should be removed",
-    );
-    assert!(
-        failures.contains_key(&blocked_ip),
-        "active block should be preserved",
-    );
-    assert!(
-        failures.contains_key(&recent_ip),
-        "in-window failure should be preserved",
-    );
-}
-
-#[test]
-fn install_token_format_short_circuits_obvious_garbage() {
-    // 32-byte hex token = 64 lowercase hex chars 才是合法格式;
-    // 任何不符合的输入应在落到 registry flock 之前就被拒掉。
-    let valid = "0123456789abcdef".repeat(4);
-    assert!(is_well_formed_install_token(&valid));
-    assert!(!is_well_formed_install_token(""));
-    assert!(!is_well_formed_install_token(&"a".repeat(63)));
-    assert!(!is_well_formed_install_token(&"a".repeat(65)));
-    // 大写不被接受 —— 与 generate_token 的 lowercase hex 输出对齐。
-    assert!(!is_well_formed_install_token(&"A".repeat(64)));
-    // 非 hex 字符
-    assert!(!is_well_formed_install_token(&"z".repeat(64)));
-}
-
-#[test]
-fn install_admission_blocks_after_repeated_failures() {
-    let controller = InstallAdmissionController::new(InstallAdmissionConfig {
-        auth_fail_window_secs: 60,
-        auth_fail_max_attempts: 2,
-        auth_block_secs: 300,
-    });
-    let client_ip: IpAddr = "198.51.100.24".parse().expect("ip");
-
-    // 阈值前应放行
-    assert!(controller.check(client_ip).is_ok());
-    controller.record_auth_failure(client_ip);
-    controller.record_auth_failure(client_ip);
-
-    match controller.check(client_ip) {
-        Err(retry_after_secs) => assert!(retry_after_secs > 0),
-        Ok(()) => panic!("client should be temporarily blocked after threshold"),
     }
 }
